Change MaxFCntGap to uint32 so it is the same type as FCnt.

+++ constants.go
@@ -10,7 +10,7 @@ const (
 	ReceiveDelay2    time.Duration = time.Second * 2
 	JoinAcceptDelay1 time.Duration = time.Second * 5
 	JoinAcceptDelay2 time.Duration = time.Second * 6
-	MaxFCntGap       uint16        = 16384
+	MaxFCntGap       uint32        = 16384
 	ADRAckLimit                    = 64
 	ADRAckDelay                    = 32
 	AckTimeoutMin    time.Duration = time.Second // AckTimeout = 2 +/- 1 (random value between 1 - 3)
